requirements.txt
appnope==0.1.0
autolink==0.1.2
backcall==0.1.0
beautifulsoup4==4.6.1
bleach==2.1.3
boto3==1.7.40
botocore==1.10.40
certifi==2018.4.16
chardet==3.0.4
decorator==4.3.0
Django==2.1
django-appconf==1.0.2
django-autolinks==0.1.0
django-bootstrap3==10.0.1
django-classy-tags==0.8.0
django-debug-toolbar==1.9.1
django-easy-maps==0.9.3
django-extensions==2.0.7
django-filter==2.0.0
django-imagekit==4.0.2
django-map-widgets==0.1.9
django-material==1.2.2
django-storages==1.6.6
djangorestframework==3.8.2
docutils==0.14
entrypoints==0.2.3
geographiclib==1.49
geopy==1.16.0
html5lib==1.0.1
idna==2.6
ipykernel==4.8.2
ipython==6.4.0
ipython-genutils==0.2.0
ipywidgets==7.2.1
jedi==0.12.0
Jinja2==2.10
jmespath==0.9.3
jsonfield==2.0.2
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
lxml==4.2.4
Markdown==2.6.11
MarkupSafe==1.0
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.5.0
numpy==1.15.0
pandas==0.23.4
pandocfilters==1.4.2
parso==0.2.1
pexpect==4.5.0
pickleshare==0.7.4
pilkit==2.0
Pillow==5.1.0
prompt-toolkit==1.0.15
psycopg2==2.7.5
psycopg2-binary==2.7.5
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.7.3
pytz==2018.4
pyzmq==17.0.0
qtconsole==4.3.1
raven==6.9.0
requests==2.19.1
s3transfer==0.1.13
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
smarturls==0.1.8
sqlparse==0.2.4
terminado==0.8.1
testpath==0.3.1
tokenizer==1.0.3
tornado==5.0.2
traitlets==4.3.2
uritools==2.2.0
urlextract==0.8.3
urllib3==1.23
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.2.1
